+++ client/Packages/com.beamable/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add email validation
+
 ### Fixed
 
 - Fixed wrong error code when providing wrong password while logging in.